+++ tests/Feature/Framework/DocBlockServiceAware/Module/Infrastructure/Persistence/FakeDoctrineEntityManager.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Framework\DocBlockServiceAware\Module\Infrastructure\Persistence;
+
+final class FakeDoctrineEntityManager
+{
+    public function findAdminName(int $id): string
+    {
+        return "fake-admin(id:{$id})";
+    }
+}